+++ src/frontend/web/src/components/Welcome/Body/WsList.tsx
@@ -1,9 +1,9 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import styled from 'styled-components';
 import expand from '../../../icon/expand.svg';
 import WsItem from './WsItem';
 import { IWorkspace } from '../../../../types/workspace.type';
-import { getWsListAPI } from '../../../Api/workspace';
+import getWsListHook from '../../../hook/getWsList';
 
 const Container = styled.section`
   border: 4px solid rgba(255, 255, 255, 0.2);
@@ -63,18 +63,7 @@ const ShowMoreIcon = styled.div`
 `;
 
 export default function WsList() {
-  const [page, setPage] = useState(0);
-  const [wsList, setWsList] = useState<IWorkspace[]>([]);
-
-  const getWsList = async () => {
-    const newWsList = await getWsListAPI(page);
-    setWsList([...wsList, ...newWsList.content]);
-    setPage((state) => state + 1);
-  };
-
-  useEffect(() => {
-    getWsList();
-  }, []);
+  const [wsList, getWsList] = getWsListHook();
 
   return (
     <Container>
