+++ src/frontend/web/src/pages/Welcome.tsx
@@ -1,5 +1,6 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import styled from 'styled-components';
+import setTitleHook from '../hook/setTitle';
 import WelcomeHeader from '../components/Welcome/Header';
 import WelcomeBody from '../components/Welcome/Body';
 
@@ -12,9 +13,7 @@ const WelcomePage = styled.div`
 `;
 
 export default function Welcome() {
-  useEffect(() => {
-    document.title = 'Welcome!';
-  }, []);
+  setTitleHook('Welcome');
 
   return (
     <WelcomePage>
